requirements.txt
# Core Dependencies
torch==1.12.1
torchvision==0.13.1
opencv-python==4.5.5.64
numpy==1.24.4

# Data Processing
pandas>=1.3.0
scipy>=1.7.0
Pillow>=8.0.0
pyyaml>=5.4.1

# Utilities
tqdm>=4.62.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0

# Development
tensorboard>=2.6.0
black>=21.7b0
flake8>=3.9.0
pytest>=6.2.0